props to resuable components
