--- src/components/ImageScreen.js
@@ -1,14 +1,0 @@
-import React from "react";
-import { View, Text, StyleSheet } from "react-native";
-
-const ImageScreen = () => {
-    return <Text style={style.textStyle}>Image Screen</Text>;
-}
-
-const style = StyleSheet.create({
-    textStyle: {
-        fontSize: 40
-    }
-});
-
-export default ImageScreen;
